refactoring can code
